+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 + Fix hidden by default is not used when rolling advantage or crits
 + Fix issue with inventory filters
 + Fix issues on FF + Linux that bricks equipping and unequipping items
++ Fix issue where scene metadata did not get updated with the newest version
 
 # 3.6.0